+++ tests/QATools/QATools/BEM/Fixture/Element/BlockChild.php
@@ -16,4 +16,22 @@ use QATools\QATools\BEM\Element\Block;
 class BlockChild extends Block
 {
 
+	/**
+	 * Example of dynamic method.
+	 *
+	 * @param string $method    Method to proxy.
+	 * @param array  $arguments Method arguments.
+	 *
+	 * @return mixed
+	 * @throws \InvalidArgumentException When given method doesn't exist.
+	 */
+	public function __call($method, array $arguments)
+	{
+		if ( $method === 'dynamicMethod' ) {
+			return 'OK';
+		}
+
+		throw new \InvalidArgumentException('The "' . $method . '" doesn\'t exist.');
+	}
+
 }