+++ tests/QATools/QATools/BEM/Proxy/ElementProxyTest.php
@@ -73,7 +73,14 @@ class ElementProxyTest extends AbstractProxyTestCase
 	 */
 	protected function createElement($replace_element_class = true)
 	{
-		return new $this->collectionClass('sample-name', $this->locator, $this->pageFactory);
+		/** @var ElementProxy $proxy */
+		$proxy = new $this->collectionClass('sample-name', $this->locator, $this->pageFactory);
+
+		if ( $replace_element_class ) {
+			$proxy->setClassName('\\tests\\QATools\\QATools\\BEM\\Fixture\\Element\\ElementChild');
+		}
+
+		return $proxy;
 	}
 
 }